\name{chi2fisher-package}
\alias{chi2fisher-package}
\alias{chi2fisher}
\docType{package}
\title{
\packageTitle{chi2fisher}
}
\description{
\packageDescription{chi2fisher}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chi2fisher}
\packageIndices{chi2fisher}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{chi2fisher}

Maintainer: \packageMaintainer{chi2fisher}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
